feat: unify the code structure of mongodb and rocksdb

+++ mongodb-implementation/src/main/kotlin/io/github/heyhey123/xiaojieorm/impl/mongo/queries/MongoDelete.kt
@@ -21,6 +21,6 @@ class MongoDelete(
         } else {
             collection.deleteMany(filter)
         }
-        return WriteResult(result.deletedCount.toInt())
+        return WriteResult(result.deletedCount)
     }
 }
